tools/requirements.txt
# FedRAMP Crypto Module Validation Tools
# Python 3.10+

# YAML/JSON processing
PyYAML>=6.0.1
ruamel.yaml>=0.18.0

# JSON Schema validation
jsonschema>=4.21.0

# Web scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# CLI utilities
click>=8.1.0
rich>=13.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Type hints
typing-extensions>=4.9.0
